拆分Line
