+++ src/manimgeo/components/__init__.py
@@ -1,5 +1,5 @@
-from ..components.base import *
+from ..components.base import GeometryAdapter, BaseGeometry
 from .point import Point
-from .line.line import *
+from .line import Line, LineSegment, Ray, InfinityLine
 from .vector.vector import *
 from .circle.circle import *
