+++ src/manimgeo/components/line/construct.py
@@ -1,0 +1,7 @@
+from typing import Union, Literal
+
+LineConstructType = Literal[
+    "PP", "PV", "TranslationLV", "VerticalPL", "ParallelPL",
+    # "TangentsCirP", "TangentsOutCirCir", "TangentsInCirCir",
+]
+Number = Union[float, int]
